InterviewAIx/requirements.txt
fastapi
uvicorn
langgraph
openai
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
